refactored out ImplicitSnitch
